+++ packages/sdk-ts/src/core/modules/exchange/msgs/MsgCreateDerivativeLimitOrder.ts
@@ -5,10 +5,7 @@ import {
   OrderTypeMap,
 } from '@injectivelabs/chain-api/injective/exchange/v1beta1/exchange_pb'
 import { MsgBase } from '../../MsgBase'
-import {
-  amountToCosmosSdkDecAmount,
-  getTriggerPrice,
-} from '../../../../utils/numbers'
+import { amountToCosmosSdkDecAmount } from '../../../../utils/numbers'
 
 export declare namespace MsgCreateDerivativeLimitOrder {
   export interface Params {
@@ -86,7 +83,9 @@ export default class MsgCreateDerivativeLimitOrder extends MsgBase<
       ...initialParams,
       price: amountToCosmosSdkDecAmount(initialParams.price).toFixed(),
       margin: amountToCosmosSdkDecAmount(initialParams.margin).toFixed(),
-      triggerPrice: getTriggerPrice(initialParams.triggerPrice),
+      triggerPrice: amountToCosmosSdkDecAmount(
+        initialParams.triggerPrice || 0,
+      ).toFixed(),
       quantity: amountToCosmosSdkDecAmount(initialParams.quantity).toFixed(),
     } as MsgCreateDerivativeLimitOrder.Params