+++ packages/sdk-ts/src/core/modules/exchange/msgs/MsgCreateSpotMarketOrder.ts
@@ -5,10 +5,7 @@ import {
   OrderTypeMap,
 } from '@injectivelabs/chain-api/injective/exchange/v1beta1/exchange_pb'
 import { MsgBase } from '../../MsgBase'
-import {
-  amountToCosmosSdkDecAmount,
-  getTriggerPrice,
-} from '../../../../utils/numbers'
+import { amountToCosmosSdkDecAmount } from '../../../../utils/numbers'
 
 export declare namespace MsgCreateSpotMarketOrder {
   export interface Params {
@@ -83,7 +80,9 @@ export default class MsgCreateSpotMarketOrder extends MsgBase<
     const params = {
       ...initialParams,
       price: amountToCosmosSdkDecAmount(initialParams.price).toFixed(),
-      triggerPrice: getTriggerPrice(initialParams.triggerPrice),
+      triggerPrice: amountToCosmosSdkDecAmount(
+        initialParams.triggerPrice || 0,
+      ).toFixed(),
       quantity: amountToCosmosSdkDecAmount(initialParams.quantity).toFixed(),
     } as MsgCreateSpotMarketOrder.Params