+++ packages/sdk-ts/src/core/modules/insurance/msgs/MsgUnderwrite.spec.ts
@@ -1,0 +1,85 @@
+import { MsgUnderwrite as BaseMsgUnderwrite } from '@injectivelabs/chain-api/injective/insurance/v1beta1/tx_pb'
+import { BigNumberInBase } from '@injectivelabs/utils'
+import MsgUnderwrite from './MsgUnderwrite'
+import { mockFactory } from '@injectivelabs/test-utils'
+import snakecaseKeys from 'snakecase-keys'
+
+const params: MsgUnderwrite['params'] = {
+  marketId: mockFactory.derivativeMarketId,
+  injectiveAddress: mockFactory.injectiveAddress,
+  amount: {
+    amount: new BigNumberInBase(1).toFixed(),
+    denom: 'inj',
+  },
+}
+
+const protoType = '/injective.insurance.v1beta1.MsgUnderwrite'
+const protoTypeAmino = 'insurance/MsgUnderwrite'
+const protoParams = {
+  marketId: params.marketId,
+  sender: params.injectiveAddress,
+  deposit: params.amount,
+}
+
+const message = MsgUnderwrite.fromJSON(params)
+
+describe.only('MsgUnderwrite', () => {
+  it('generates proper proto', () => {
+    const proto = message.toProto()
+
+    expect(proto instanceof BaseMsgUnderwrite).toBe(true)
+    expect(proto.toObject()).toStrictEqual(protoParams)
+  })
+
+  it('generates proper data', () => {
+    const data = message.toData()
+
+    expect(data).toStrictEqual({
+      '@type': protoType,
+      ...protoParams,
+    })
+  })
+
+  it('generates proper amino', () => {
+    const amino = message.toAmino()
+
+    expect(amino).toStrictEqual({
+      type: protoTypeAmino,
+      ...protoParams,
+    })
+  })
+
+  it('generates proper Eip712 types', () => {
+    const eip712Types = message.toEip712Types()
+
+    expect(Object.fromEntries(eip712Types)).toStrictEqual({
+      TypeDeposit: [
+        { name: 'denom', type: 'string' },
+        { name: 'amount', type: 'string' },
+      ],
+      MsgValue: [
+        { name: 'sender', type: 'string' },
+        { name: 'market_id', type: 'string' },
+        { name: 'deposit', type: 'TypeDeposit' },
+      ],
+    })
+  })
+
+  it('generates proper Eip712 values', () => {
+    const eip712 = message.toEip712()
+
+    expect(eip712).toStrictEqual({
+      type: protoTypeAmino,
+      value: snakecaseKeys(protoParams),
+    })
+  })
+
+  it('generates proper web3', () => {
+    const web3 = message.toWeb3()
+
+    expect(web3).toStrictEqual({
+      '@type': protoType,
+      ...protoParams,
+    })
+  })
+})