+++ packages/contracts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.30](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.29...@injectivelabs/contracts@1.0.30) (2022-09-14)
+
+**Note:** Version bump only for package @injectivelabs/contracts
+
+
+
+
+
 ## [1.0.29](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/contracts@1.0.28...@injectivelabs/contracts@1.0.29) (2022-09-13)
 
 **Note:** Version bump only for package @injectivelabs/contracts