+++ packages/exceptions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.11...@injectivelabs/exceptions@1.0.12) (2022-09-14)
+
+**Note:** Version bump only for package @injectivelabs/exceptions
+
+
+
+
+
 ## [1.0.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.10...@injectivelabs/exceptions@1.0.11) (2022-09-13)
 
 **Note:** Version bump only for package @injectivelabs/exceptions